backend/requirements/development.txt
# Dependencias para desarrollo
-r base.txt

# Debug y desarrollo
django-debug-toolbar==4.4.6
django-extensions==3.2.3

# Testing
pytest==8.3.3
pytest-django==4.9.0
pytest-cov==6.0.0
factory-boy==3.3.1

# Code quality
flake8==7.1.1
black==24.10.0
isort==5.13.2
mypy==1.13.0

# Documentation
Sphinx==8.1.3
